// src/lib/api/models/getApiCustomersType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export type GetApiCustomersType = typeof GetApiCustomersType[keyof typeof GetApiCustomersType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiCustomersType = {
  BUYER: 'BUYER',
  RENTER: 'RENTER',
  INVESTOR: 'INVESTOR',
  SELLER: 'SELLER',
  LANDLORD: 'LANDLORD',
} as const;