// src/lib/api/models/getApiCustomersStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export type GetApiCustomersStatus = typeof GetApiCustomersStatus[keyof typeof GetApiCustomersStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetApiCustomersStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  BLACKLISTED: 'BLACKLISTED',
} as const;